move config to dedicated module
