+++ src/drf_pydantic/__init__.py
@@ -3,4 +3,4 @@ __all__ = [
 ]
 __version__ = "2.7.0"
 
-from .base_model import BaseModel
+from drf_pydantic.base_model import BaseModel
